vastly untested refactor. pushing for backup purposes
