+++ src/main/res/menu/fragment_conversation.xml
@@ -1,12 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_add"
-        android:icon="?attr/icon_new"
-        android:orderInCategory="10"
-        app:showAsAction="always"
-        android:title="@string/action_add"/>
+      xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
         android:id="@+id/action_security"
         android:icon="?attr/icon_not_secure"
@@ -103,16 +97,4 @@
         android:orderInCategory="71"
         app:showAsAction="never"
         android:title="@string/enable_notifications"/>
-
-    <item
-        android:id="@+id/action_accounts"
-        android:orderInCategory="90"
-        app:showAsAction="never"
-        android:title="@string/action_accounts"/>
-    <item
-        android:id="@+id/action_settings"
-        android:orderInCategory="100"
-        app:showAsAction="never"
-        android:title="@string/action_settings"/>
-
 </menu>
